:sparkles: [Feat] client-side 확인하는 CAN_USE_DOM 유틸 추가및 Quote type 정리
