+++ src/types/mover.ts
@@ -52,18 +52,6 @@ export interface RequestDetails {
   cost: number;
 }
 
-export interface QuoteDetailsData {
-  id: number;
-  requestDate: string;
-  service: number;
-  isDesignated: boolean;
-  isConfirmed?: boolean;
-  name: string;
-  movingDate: string;
-  pickupAddress: string;
-  dropOffAddress: string;
-}
-
 interface UserBaseData {
   id: number;
   name: string;
